Split out functions into GuitarCommon
