+++ Sources/Guitar.swift
@@ -84,28 +84,6 @@ public struct Guitar {
     }
 }
 
-// MARK: - Common Evaluations and Tests
-
-public extension Guitar {
-    /// Tests a string to check if it is a valid email address by using a regular expression.
-    ///
-    /// - Parameters:
-    ///     - email: The string that needs to be evaluated.
-    ///
-    /// - Returns: `true` if `string` is a valid email address, otherwise `false`.
-    static func isValidEmail(email: String) -> Bool {
-        return Guitar(chord: .email).test(string: email)
-    }
-
-    /// Sanitizes of a string by removing all non-Alphanumeric characters (excluding whitespaces)
-    ///
-    /// - Parameter string: The string that should be sanitized.
-    /// - Returns: The sanitized string.
-    static func sanitze(string: String) -> String {
-        return Guitar(chord: .nonAlphanumeric).replaceOccurences(in: string, with: " ").trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-}
-
 // MARK: - Helpers
 
 public extension Guitar {
